Fix memory leak in the activation process
